requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
python-dotenv==1.0.0
Werkzeug==2.3.3
